Finished delete, error handling for delete.

+++ routes/note.controller.js
@@ -8,6 +8,15 @@ const noteRepository = AppDataSource.getRepository(Note);
 // Create new record
 router.post("/notes", async (req, res) => {
   const { title, body } = req.body;
+  if (title === "") {
+    console.log("Note title can't be empty.");
+    return res.status(400).json({ error: "Unable to create note" });
+  }
+  if (body === "") {
+    console.log("Note body can't be empty.");
+    return res.status(400).json({ error: "Unable to create note" });
+  }
+
   const note = new Note();
   note.title = title;
   note.body = body;
@@ -45,6 +54,15 @@ router.put("/notes/:id", async (req, res) => {
   const { id } = req.params;
   const { title, body } = req.body;
 
+  if (title === "") {
+    console.log("Note title can't be empty.");
+    return res.status(400).json({ error: "Unable to create note" });
+  }
+  if (body === "") {
+    console.log("Note body can't be empty.");
+    return res.status(400).json({ error: "Unable to create note" });
+  }
+
   try {
     await noteRepository.update(id, { title, body });
     const updatedNote = await noteRepository.findOne(id);
@@ -58,12 +76,29 @@ router.put("/notes/:id", async (req, res) => {
 router.delete("/notes/:id", async (req, res) => {
   const { id } = req.params;
 
+  if (id === null) {
+    console.log("Must include ID in search.");
+    return res.status(400).json({ error: "ID missing." });
+  }
   try {
-    await noteRepository.delete(id);
-    console.log(`Note ${id} deleted.`);
-    return res.status(204).send();
-  } catch (error) {
-    return res.status(500).json({ error: "Unable to delete note" });
+    const note = await noteRepository.findOne(id);
+
+    if (note) {
+      try {
+        // if id exists
+        await noteRepository.delete(id);
+        console.log(`Note ${id} deleted.`);
+        // return 200
+
+        // else return 400
+        return res.status(200).send();
+      } catch (error) {
+        return res.status(500).json({ error: "Unable to delete note" });
+      }
+    }
+  } catch {
+    console.log(`Note not found with id ${id}`);
+    return res.status(404).json({ error: "Note not found" });
   }
 });
